device: re-read initial hw-address before activating connection
Previously, we would only once read the initial-hw address during device realization.

When a device activates, NetworkManager always sets the MAC address as configured
in the cloned-mac-address setting -- or, if unspecified -- it falls back
to use the permanent hardware-address instead.

Currently, NetworkManager does not support any other scheme, like using
randomized MAC addresses, stable-private MAC addresses or simply not
touch the MAC address when activating a device.

The latter, not touching the MAC address might not make too much sense
either, because the MAC address is in an undefined state, especially
after multiple UP/DOWN cylces.

Anyway, when deactivating the device, we reset the MAC address to the
"inital MAC address".

This patch changes, that the "initial MAC address" is re-read every time
before activating the device, contrary to reading it once in the
beginning.

https://bugzilla.gnome.org/show_bug.cgi?id=708820