device: reconfigure IP addressing after bringing up device
For changing the hardware address, we must bring the device down. When doing
that, IP addressing is lost and it must be re-configured after bringing the
device up again.

We already do something similar in device_link_changed(), but that might
not be sufficient, because device_link_changed() is run on an idle
handler, thus, while chaning the hardware address it has no chance to
run (or see that the device was shortly down).

https://bugzilla.redhat.com/show_bug.cgi?id=1309899